core: Move FontSet to font/font_set.rs
This makes font.rs smaller, as it's a pretty big file.
